[Group 21] req11

+++ back-end/src/database/migrations/20230120190839-products.js
@@ -6,7 +6,7 @@ module.exports = {
     await queryInterface.createTable('products', {
       id: { type: Sequelize.INTEGER, primaryKey: true, autoIncrement: true, allowNull: false },
       name: { type: Sequelize.STRING, allowNull: false, unique: true },
-      price: { type: Sequelize.FLOAT, allowNull: false },
+      price: { type: Sequelize.DECIMAL(9, 2), allowNull: false },
       urlImage: { type: Sequelize.STRING, allowNull: false, defaultValue: '', field: 'url_image' },
     });
   },
